requirements.txt
flask==2.2.3
Flask-SocketIO==5.3.3
pandas==1.5.3
plotly==5.14.0
google-api-python-client==2.83.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
tabulate==0.9.0
paho-mqtt==1.6.1
